Add eodms-api-client to data discovery tools
Shameless self-plug. As the userbase and SAR data archives (including RADARSAT-1, RADARSAT-2 and RADARSAT CONSTELLATION MISSION) grow on EODMS it may be useful for some people to have a programmatic way to run queries and related workflows.

This package was designed to have a similar user experience as the excellent `sentinelsat` package.